backend/pyproject.toml
[project]
name = "sapti-backend"
version = "0.1.0"
description = "Sapti AI — The Hive Mind Protocol. An evolving AI companion powered by 7 LangGraph agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.6",
    "fastapi>=0.135.2",
    "google-genai>=1.70.0",
    "httpx>=0.28.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langgraph>=1.1.3",
    "litellm>=1.82.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "pyyaml>=6.0.3",
    "sse-starlette>=3.3.4",
    "structlog>=25.5.0",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.42.0",
]
